requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
